backend/requirements.txt
# ---- Core API ----
fastapi==0.110.3
starlette==0.37.2
uvicorn==0.30.1

# ---- Validation ----
pydantic==2.7.4
pydantic-settings==2.3.4

# ---- Async ----
anyio==4.4.0
h11==0.14.0

# ---- Utilities ----
python-dotenv==1.0.1
click==8.1.7
typing-extensions==4.12.2

# ---- Presidio ----
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# ---- NLP (LOCKED + COMPATIBLE) ----
spacy==3.7.4
thinc==8.2.5
blis==0.7.11
numpy==1.26.4
typer==0.9.4
weasel==0.3.4
langcodes==3.5.1
cloudpathlib==0.16.0
smart-open==6.4.0

# ---- Transformers & Embeddings ----
torch
torchvision
torchaudio
transformers
sentence-transformers

# ---- Vector Search ----
faiss-cpu

# ---- ML Models ----
scikit-learn
joblib
lightgbm

# ---- Language Detection ----
langdetect
